update: synthesiser cleanup
